backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==9.3.0
PyJWT==2.1.0
requests==2.26.0
django-filter==23.1
djoser
django-cors-headers==4.5.0
psycopg2-binary==2.9.3